edits

+++ Tools/Coyote/Cli/CommandLineParser.cs
@@ -348,7 +348,7 @@ namespace Microsoft.Coyote.Cli
                 Arity = ArgumentArity.Zero
             };
 
-            var seedOption = new Option<int>(
+            var seedOption = new Option<uint>(
                 name: "--seed",
                 description: "Specify the random value generator seed.")
             {
@@ -1010,7 +1010,7 @@ namespace Microsoft.Coyote.Cli
                         this.Configuration.IsPartialOrderSamplingEnabled = true;
                         break;
                     case "seed":
-                        this.Configuration.RandomGeneratorSeed = (uint)result.GetValueOrDefault<int>();
+                        this.Configuration.RandomGeneratorSeed = result.GetValueOrDefault<uint>();
                         break;
                     case "liveness-temperature-threshold":
                         this.Configuration.LivenessTemperatureThreshold = result.GetValueOrDefault<int>();
